explode functions
